build: split packages
