+++ legacy-connectors/NOTICE.txt
