--- packaging/src/main/assemblies/filter.properties
@@ -1,4 +1,0 @@
-product.name=Neo4j Connector for Confluent
-product.description=The connector runs on Apache Kafka Connect Clusters to sink data in Kafka topics into Neo4j and Aura databases and can use Neo4j/Aura as a source for Kafka topics via Cypher templated queries.
-product.docs.url=https://neo4j.com/docs
-product.summary=Support, including for Custom Connectors, through <a href=\"https://neo4j.com/\">Neo4j</a>
