--- packaging/src/main/assemblies/kafka-jar.xml
@@ -1,35 +1,0 @@
-<assembly xmlns="http://maven.apache.org/ASSEMBLY/2.2.0"
-          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-          xsi:schemaLocation="http://maven.apache.org/ASSEMBLY/2.2.0 https://maven.apache.org/xsd/assembly-2.2.0.xsd">
-    <id>kc-oss-jar</id>
-
-    <formats>
-        <format>jar</format>
-    </formats>
-
-    <includeBaseDirectory>false</includeBaseDirectory>
-
-    <fileSets>
-        <fileSet>
-            <directory>src/main/distributions/text/doc</directory>
-            <outputDirectory>META-INF</outputDirectory>
-            <lineEnding>keep</lineEnding>
-            <filtered>true</filtered>
-            <excludes>
-                <exclude>**/.keep</exclude>
-                 <exclude>**/*.formatted</exclude>
-            </excludes>
-            <directoryMode>0755</directoryMode>
-            <fileMode>0644</fileMode>
-        </fileSet>
-    </fileSets>
-
-    <dependencySets>
-        <dependencySet>
-            <outputDirectory>/</outputDirectory>
-            <useProjectArtifact>false</useProjectArtifact>
-            <unpack>true</unpack>
-            <scope>runtime</scope>
-        </dependencySet>
-    </dependencySets>
-</assembly>
