+++ packaging/src/main/assemblies/kafka-legacy-zip.xml
@@ -1,0 +1,67 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<assembly xmlns="http://maven.apache.org/ASSEMBLY/2.2.0"
+          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+          xsi:schemaLocation="http://maven.apache.org/ASSEMBLY/2.2.0 http://maven.apache.org/xsd/assembly-2.2.0.xsd">
+    <id>kc-oss-legacy-zip</id>
+
+    <formats>
+        <format>zip</format>
+    </formats>
+
+    <fileSets>
+        <fileSet>
+            <directory>src/main/distributions/text</directory>
+            <outputDirectory/>
+            <lineEnding>keep</lineEnding>
+            <filtered>true</filtered>
+            <excludes>
+                <exclude>**/.keep</exclude>
+                <exclude>**/*.formatted</exclude>
+            </excludes>
+            <directoryMode>0755</directoryMode>
+            <fileMode>0644</fileMode>
+        </fileSet>
+    </fileSets>
+
+    <dependencySets>
+        <dependencySet>
+            <outputDirectory>lib</outputDirectory>
+            <outputFileNameMapping>${artifact.artifactId}-${artifact.version}.${artifact.extension}
+            </outputFileNameMapping>
+            <directoryMode>0755</directoryMode>
+            <fileMode>0755</fileMode>
+            <unpack>false</unpack>
+            <useProjectArtifact>false</useProjectArtifact>
+            <useTransitiveDependencies>true</useTransitiveDependencies>
+            <includes>
+                <include>
+                    org.neo4j.connectors.kafka:legacy-connectors
+                </include>
+                <include>
+                    org.neo4j.connectors.kafka:common
+                </include>
+                <include>
+                    org.neo4j.connectors.kafka:sink
+                </include>
+                <include>
+                    org.neo4j.connectors.kafka:source
+                </include>
+            </includes>
+        </dependencySet>
+        <dependencySet>
+            <outputDirectory>lib</outputDirectory>
+            <outputFileNameMapping>${artifact.artifactId}-${artifact.version}.${artifact.extension}
+            </outputFileNameMapping>
+            <directoryMode>0755</directoryMode>
+            <fileMode>0755</fileMode>
+            <unpack>false</unpack>
+            <useProjectArtifact>false</useProjectArtifact>
+            <useTransitiveDependencies>true</useTransitiveDependencies>
+            <excludes>
+                <exclude>
+                    org.neo4j.connectors.kafka:*
+                </exclude>
+            </excludes>
+        </dependencySet>
+    </dependencySets>
+</assembly>
