+++ sink-connector/LICENSES.txt
