+++ sink-connector/NOTICE.txt
