+++ sink/LICENSES.txt
@@ -6,7 +6,6 @@ libraries. For an overview of the licenses see the NOTICE.txt file.
 Apache Software License, Version 2.0
   Apache Commons Collections
   Apache Commons Lang
-  common
   Jackson-annotations
   Jackson-core
   jackson-databind
@@ -18,7 +17,6 @@ Apache Software License, Version 2.0
   Kotlin Stdlib Jdk7
   Kotlin Stdlib Jdk8
   kotlinx-coroutines-core
-  Neo4j Java Driver
   Non-Blocking Reactive Foundation for the JVM
 ------------------------------------------------------------------------------
