+++ sink/NOTICE.txt
@@ -21,7 +21,6 @@ Third-party licenses
 Apache Software License, Version 2.0
   Apache Commons Collections
   Apache Commons Lang
-  common
   Jackson-annotations
   Jackson-core
   jackson-databind
@@ -33,7 +32,6 @@ Apache Software License, Version 2.0
   Kotlin Stdlib Jdk7
   Kotlin Stdlib Jdk8
   kotlinx-coroutines-core
-  Neo4j Java Driver
   Non-Blocking Reactive Foundation for the JVM
 
 BSD License
