+++ source-connector/LICENSES.txt
