+++ source-connector/NOTICE.txt
